Fix docoffset count
If compile_results() excluded some results from the returned list,
docoffset was only partially incremented meaning some of the same
documents would be re-queried. If compile_results() excluded everything,
then docoffset was not incremented at all, and the process got stuck.

This increments docoffset by the number of results that were processed,
so we are processing a whole page each time, and will not get stuck.